+++ CHANGELOG.md
@@ -100,7 +100,7 @@ dev-master
        - Created base class for query segment and operation: `Queries\Common\Join\Base`.
        - Updated `Segments\Segment[Walker|Visitor]` and `Operation\Visitor` to match new structure.
        - Hence updated `Providers\Traversable\ScopeEvaluator` and `Providers\Collection\OperationEvaluator`.
- - Removed obsolete query providers (`Loadable`, `Caching`).
+ - Removed obsolete query providers (`Loadable`, `Caching`) in favour of a new helper class `Providers\Utilities\QueryResultCollection`
  - Implemented new DSL query provider under `Providers\DSL`.
  - New structure of query providers
     - `RepositoryProvider` decorates the `QueryProvider`